Adjusting folder structure
