+++ src/tasks/CombineTask.js
