+++ src/tasks/GulpPaths.js
