+++ src/tasks/JavaScriptTask.js
