+++ src/tasks/recipes/phpunit.js
@@ -1,4 +1,5 @@
-import TestingTask from './conductors/TestingTask';
+import TestingTask from '../TestingTask';
+
 /*
  |----------------------------------------------------------------
  | PHPUnit Testing
